updated weight for cosebi calculation, fixed bug with theta integration support

+++ input/arbitrary_summary/script_weights/get_weights_cosebis.py
@@ -5,18 +5,24 @@ from mpmath import mp
 import mpmath
 from scipy.interpolate import interp1d
 import multiprocessing as mpi
+from scipy.signal import argrelextrema
+
+
 mpi.set_start_method("fork")
 
 mp.dps = 160
 num_cores = 8 #number of cores used
 #define constants
-Nmax = 5 # maximum mode
+Nmax = 5 # maximum COSEBI mode
 tmin = 0.5 #theta_min in arcmin
 tmax = 300.0 #theta_max in armin
+ell_min = 1 # Minimum multipole
+ell_max = 1e5 # Maximum multipole
+N_ell = int(1e5) # Number of Fourier modes
+get_W_ell_as_well = True # If true the Well are calculated
+
+#####################
 zmax = mp.log(tmax/tmin)
-ell_min = 1
-ell_max = 1e5
-N_ell = int(1e4)
 ell = np.geomspace(ell_min, ell_max, N_ell)
 
 
@@ -120,15 +126,23 @@ def tplus(tmin,tmax,n,norm,root,ntheta=10000):
 
 def tminus(tmin,tmax,n,norm,root,tp,ntheta=10000):
     tplus_func=interp1d(np.log(tp[:,0]/tmin),tp[:,1])
-    tminus = np.zeros_like(tp)
-    tminus[:,0] = theta
-    tminus[:,1]=tplus_func(z)
+    rtminus = np.zeros_like(tp)
+    rtminus[:,0] = theta
+    rtminus[:,1]= tplus_func(z)
     for i_z, val_z in enumerate(z[1:]):
         y = np.linspace(z[0],val_z, 1000)
         integrand = tplus_func(y)*(np.exp(2*(y-val_z)) - 3*np.exp(4*(y-val_z)))
-        lev.init_integral(y,integrand[:,None],True,True)
-        tminus[i_z, 1] +=  4*lev.cquad_integrate_single_well([y[0],y[-1]],0)[0]
-    return tminus
+        limits_at_mode = np.array(y[argrelextrema(integrand, np.less)[0][:]])
+        limits_at_mode_append = np.zeros(len(limits_at_mode) + 2)
+        if len(limits_at_mode) != 0:
+            limits_at_mode_append[1:-1] = limits_at_mode
+        limits_at_mode_append[0] = y[0]
+        limits_at_mode_append[-1] = y[-1]
+        
+        lev.init_integral(y,integrand[:,None],False,False)
+        rtminus[i_z, 1] +=  4*lev.cquad_integrate_single_well(limits_at_mode_append,0)[0]
+        #rtminus[i_z, 1] += 4*np.trapz(integrand,y)
+    return rtminus
 
 
 for nn in range(1,Nmax+1):
@@ -138,35 +152,43 @@ for nn in range(1,Nmax+1):
     lev = levin.Levin(0, 16, 32, 1e-8, 200)
     theta = tpn[:,0]
     z=np.log(theta/tmin)
-    lev.init_w_ell(z, np.ones_like(theta)[:,None])
-    tmn = tminus(tmin,tmax,1,Nn[n],rn[n], tpn)[:,1]
+    wide_theta = np.linspace(theta[0]*0.9, theta[-1]*1.1,int(1e4))
+    lev.init_w_ell(np.log(wide_theta/tmin), np.ones_like(wide_theta)[:,None])
+    tmn = tminus(tmin,tmax,1,Nn[n],rn[n], tpn)
     lev_w = levin.Levin(0, 16, 32, 1e-8, 200)
     arcmintorad = np.pi/180/60
     lev_w.init_integral(theta*arcmintorad,(theta*arcmintorad*tpn[:,1])[:,None],True,True)
+    if get_W_ell_as_well:
+        global getWell
 
-    global getWell
-
-    def getWell(ells):
-        result = lev_w.single_bessel(ells,0,theta[0]*arcmintorad,theta[-1]*arcmintorad)[0]
-        return result
-    
-    
-    pool = mpi.Pool(num_cores)
-    result_Well = np.array(pool.map(
-                    getWell, ell))
-    pool.close()
-    pool.terminate()
+        def getWell(ells):
+            result = lev_w.single_bessel(ells,0,theta[0]*arcmintorad,theta[-1]*arcmintorad)[0]
+            return result
+        
+        
+        pool = mpi.Pool(num_cores)
+        result_Well = np.array(pool.map(
+                        getWell, ell))
+        pool.close()
+        pool.terminate()
+    tpn[:,1] /= 2
+    tmn[:,1] /= 2 
+    filenamep = "./../Tplus"+str(nn)+"_0.50-300.00.table"
+    filenamem = "./../Tminus"+str(nn)+"_0.50-300.00.table"
+    marika_tp = np.loadtxt(filenamep)
+    marika_tm = np.loadtxt(filenamem)
     if nn < 10:
-        file_tpn = "./../Tplus_"+str(nn) + "_0" +str(tmin) + "_to_" + str(tmax) + ".table"
-        file_tmn = "./../Tminus_"+str(nn) + "_0" +str(tmin) + "_to_" + str(tmax) + ".table"
-        file_Wn = "./../WnLog_"+str(nn) + "_0" +str(tmin) + "_to_" + str(tmax) + ".table"
+        file_tpn = "./../Tp_0"+str(nn) + "_" +str(tmin) + "_to_" + str(tmax) + ".table"
+        file_tmn = "./../Tm_0"+str(nn) + "_" +str(tmin) + "_to_" + str(tmax) + ".table"
+        file_Wn = "./../Wn_0"+str(nn) + "_" +str(tmin) + "_to_" + str(tmax) + ".table"
     else:
-        file_tpn = "./../Tplus_"+str(nn) + "_" +str(tmin) + "_to_" + str(tmax) + ".table"
-        file_tmn = "./../Tminus_"+str(nn) + "_" +str(tmin) + "_to_" + str(tmax) + ".table"
-        file_Wn = "./../WnLog_"+str(nn) + "_" +str(tmin) + "_to_" + str(tmax) + ".table"
-    np.savetxt(file_tpn,2*tpn)
-    np.savetxt(file_tmn,2*tmn)    
-    np.savetxt(file_Wn, np.array([ell,result_Well]).T)
+        file_tpn = "./../Tp_"+str(nn) + "_" +str(tmin) + "_to_" + str(tmax) + ".table"
+        file_tmn = "./../Tm_"+str(nn) + "_" +str(tmin) + "_to_" + str(tmax) + ".table"
+        file_Wn = "./../Wn_"+str(nn) + "_" +str(tmin) + "_to_" + str(tmax) + ".table"
+    np.savetxt(file_tpn,tpn)
+    np.savetxt(file_tmn,tmn)    
+    if get_W_ell_as_well:
+        np.savetxt(file_Wn, np.array([ell,result_Well]).T)
